requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
pydantic[email]
python-multipart
pyodbc
python-jose[cryptography]
passlib[argon2]
argon2-cffi
gunicorn
slowapi
openai-whisper